renamed gcp vertex to gcp vertex gemini
