+++ gateway/src/inference/providers/gcp_vertex_gemini.rs
