+++ gateway/tests/e2e/providers/gcp_vertex_gemini.rs
